gbsjuridico-backend/requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cryptography==37.0.1
distlib==0.3.4
et-xmlfile==1.1.0
filelock==3.7.1
Flask==2.2.2
Flask-Cors==3.0.10
greenlet==1.1.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.8.0
MarkupSafe==2.1.2
MouseInfo==0.1.3
mysql-connector-python==8.0.32
openpyxl==3.0.9
outcome==1.1.0
Pillow==9.1.0
platformdirs==2.5.2
PyAutoGUI==0.9.53
pyclip==0.6.0
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMuPDF==1.19.6
pyodbc==4.0.32
pyOpenSSL==22.0.0
PyPDF2==1.27.12
pyperclip==1.8.2
pyperclip3==0.4.1
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-docx==0.8.11
python-dotenv==0.21.1
pytweening==1.0.4
pywin32==304
requests==2.27.1
selenium==4.1.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.2.2
wsproto==1.1.0
